invoice-app/server/nlp_service/requirements.txt
flask>=2.2
spacy>=3.5
python-dateutil>=2.8
